Move NavigationSettings to containers
