+++ front/app/containers/Admin/pagesAndMenu/containers/NavigationSettings/NavbarItemRow/NavbarItemRow.test.tsx
