+++ front/app/containers/Admin/pagesAndMenu/containers/NavigationSettings/VisibleNavbarItemList/VisibleNavbarItemList.test.tsx
